pyproject.toml
[tool.poetry]
name = "grocery-store"
version = "0.1.0"
description = ""
authors = ["Oleg Smirnov <oleg.smirnoff505@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
